+++ src/snowflake/snowpark/functions.py
@@ -7713,6 +7713,7 @@ def pandas_udtf(
     external_access_integrations: Optional[List[str]] = None,
     secrets: Optional[Dict[str, str]] = None,
     immutable: bool = False,
+    max_batch_size: Optional[int] = None,
 ) -> Union[UserDefinedTableFunction, functools.partial]:
     """Registers a Python class as a vectorized Python UDTF and returns the UDTF.
 
@@ -7817,6 +7818,7 @@ def pandas_udtf(
             external_access_integrations=external_access_integrations,
             secrets=secrets,
             immutable=immutable,
+            max_batch_size=max_batch_size,
         )
     else:
         return session.udtf.register(
@@ -7838,6 +7840,7 @@ def pandas_udtf(
             external_access_integrations=external_access_integrations,
             secrets=secrets,
             immutable=immutable,
+            max_batch_size=max_batch_size,
         )